pyproject.toml
[tool.poetry]
name = "anjani"
version = "2.14.3"
description = "Telegram group management bot"
license = "GPL-3.0-or-later"
authors = [
    "Gaung Ramadhan <hi@mrmiss.dev>",
    "Adek Maulana <adekzmaulana@gmail.com>",
]
readme = "README.md"
homepage = "https://userbotindo.com/anjani"
repository = "https://github.com/userbotindo/anjani"
documentation = "https://userbotindo.com/anjani/docs/home"
keywords = ["chatbot", "bot", "python", "telegram-bot", "pyrogram"]
classifiers = [
    "Environment :: Console",
    "Framework :: AsyncIO",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Communications",
    "Topic :: Communications :: Chat",
]

[tool.poetry.scripts]
anjani = "anjani.main:start"

[tool.poetry.dependencies]
python = "~=3.9"
aiohttp = "^3.8.4"
aiopath = [
    { version = "^0.5.12", python = "<3.10" },
    { version = "^0.6.11", python = ">=3.10" },
]
aiorun = ">=2022.11.1,<2024.0.0"
certifi = { version = ">=2022.12.7,<2025.0.0", platform = "windows" }
colorlog = "^6.7.0"
frozenlist = "^1.3.3"
meval = "^2.5"
multidict = "^6.0.4"
pymongo = "^4.3.3"
pyrofork = "^2.3.13"
python-dotenv = ">=0.21.1,<1.1.0"
PyYAML = "^6.0"
TgCrypto = "^1.2.5"
typing-extensions = "^4.5.0"
uvloop = { version = ">=0.17,<0.20", optional = true, platform = "linux" }
yarl = "^1.8.2"
aiocache = "^0.12.0"
prometheus-client = "^0.20.0"
mysql-connector-python = "^8.4.0"

[tool.poetry.extras]
all = ["uvloop"]
uvloop = ["uvloop"]

[tool.poetry.group.dev.dependencies]
black = ">=22.12,<25.0"
isort = "^5.12.0"
pre-commit = ">=2.21,<4.0"
pylint = ">=2.16.2,<4.0.0"
flake8 = ">=5.0.4,<8.0.0"
pytest = ">=7.2.1,<9.0.0"
pytest-asyncio = ">=0.20.2,<0.24.0"

#
# Black config
#
[tool.black]
line-length = 100
target-version = ["py38"]

#
# Isort Config
#
[tool.isort]
profile = "black"
known_third_party = [
    "aiohttp",
    "aiopath",
    "aiorun",
    "colorlog",
    "dotenv",
    "pymongo",
    "pyrogram",
    "scikit-learn",
    "yaml",
]

#
# Pylint Config
#
[tool.pylint]
ignore = "anjani/util/db"

[tool.pylint.messages_control]
disable = [
    "attribute-defined-outside-init",
    "too-few-public-methods",
    "missing-class-docstring",
    "logging-fstring-interpolation",
    "missing-function-docstring",
]

[tool.pylint.reports]
output-format = "colorized"

[tool.pylint.logging]
logging-format-style = "new"

[tool.pylint.format]
max-line-length = 100

[tool.pylint.design]
max-args = 15

# Semantic Release Config
[tool.semantic_release]
version_variable = "anjani/__init__.py:__version__"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
